+++ R/prepInputs.R
@@ -5,7 +5,6 @@ if (getRversion() >= "3.1.0") {
 #' Download and optionally post-process files
 #'
 #' \if{html}{\figure{lifecycle-maturing.svg}{options: alt="maturing"}}
-#' \if{latex}{\figure{lifecycle-maturing.svg}{options: width=0.5in}}
 #'
 #' This function can be used to prepare R objects from remote or local data sources.
 #' The object of this function is to provide a reproducible version of
@@ -60,11 +59,11 @@ if (getRversion() >= "3.1.0") {
 #'   If \code{rasterToMatch} or \code{studyArea} are used, then this will
 #'   trigger several subsequent functions, specifically the sequence,
 #'   \emph{Crop, reproject, mask}, which appears to be a common sequence in
-#'   spatial simulation. See \code{\link{postProcess.spatialObjects}}.
+#'   spatial simulation. See \code{\link{postProcess.spatialClasses}}.
 #'
 #'   \emph{Understanding various combinations of \code{rasterToMatch}
 #'   and/or \code{studyArea}:}
-#'   Please see \code{\link{postProcess.spatialObjects}}.
+#'   Please see \code{\link{postProcess.spatialClasses}}.
 #'  }
 #'
 #' @section \code{purge}:
@@ -787,8 +786,6 @@ extractFromArchive <- function(archive,
   list(moduleName = moduleName, modulePath = modulePath, checkSums = checkSums)
 }
 
-#' @keywords internal
-.checkSumsMem <- memoise::memoise(.checkSums)
 
 .isArchive <- function(filename) {
   if (!is.null(filename)) {