+++ R/reproducible-package.R
@@ -2,7 +2,6 @@
 #'
 #' @description
 #' \if{html}{\figure{lifecycle-maturing.svg}{options: alt="maturing"}}
-#' \if{latex}{\figure{lifecycle-maturing.svg}{options: width=0.5in}}
 #'
 #' This package aims at making
 #' high-level, robust, machine and OS independent tools for making deeply