+++ NEWS.md
@@ -1,16 +1,17 @@
 Known issues: https://github.com/PredictiveEcology/reproducible/issues
 
-version 1.1.1.9000
+version 1.2.0.9000
 ==============
 
 ## New features
+* updated digesting of `Raster` class objects to account for more of the metadata (including the colortable). This will change the digest value of all `Raster` layers, causing re-run of `Cache`
 * removed `Require`, `pkgDep`, `trimVersionNumber`, `normPath`, `checkPath` that were moved to `Require` package. For backwards compatibility, these are imported and reexported
 * address permanently or temporarily new changes in GDAL>3 and PROJ>6 in the spatial packages.
 * new function `file.move` used to rename/copy files across disks (a situation where `file.rename` would fail)
 * all `DBI` type functions now have default `cachePath` of `getOption("reproducible.cachePath")`
 
 ## Dependency changes
-* net reduction in number of packages that are imported from by 6. Removed completely: `R.utils`, `tools`, `remotes`, `versions`; moved to Suggests: `rgdal`, `gdalUtils`, `googledrive`, `fastdigest`, `testthat`; added: `Require`, `lifecycle`, `usethis`
+* net reduction in number of packages that are imported from by 9. Removed completely: `memoise`, `quickPlot`, `R.utils`, `remotes`, `tools`, and `versions`; moved to Suggests: `gdalUtils`, `googledrive`, `fastdigest`, `rgdal`, `sf`, `testthat`; added: `Require`
 
 ## bug fixes
 * fix over-wide tables in PDF manual (#144)